requirements.txt
schedule==1.2.0
python-dotenv==1.0.0
PyMySQL==1.0.3
pyTelegramBotAPI==4.10.0
